% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwdsb-package.R
\docType{package}
\name{hwdsb-package}
\alias{hwdsb}
\alias{hwdsb-package}
\title{hwdsb: Collection of Functions for Hamilton-Wentworth District School Board}
\description{
These functions are to be used for data cleaning and management, data visualization, and report-building for school- and student-level data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkanters/hwdsb}
  \item Report bugs at \url{https://github.com/dkanters/hwdsb/issues}
}

}
\author{
\strong{Maintainer}: David Kanters \email{dkanters@hwdsb.on.ca}

}
\keyword{internal}
