% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwdsb_ggplot_theme.R
\name{hwdsb_ggplot_theme}
\alias{hwdsb_ggplot_theme}
\title{hwdsb_ggplot_theme}
\usage{
hwdsb_ggplot_theme()
}
\value{
a ggplot theme
}
\description{
hwdsb_ggplot_theme
}
\examples{
#mtcars |>
#  head(3) |>
#  tibble::rownames_to_column() |>
#  ggplot2::ggplot(ggplot2::aes(x = rowname, y = mpg)) +
#  ggplot2::geom_col() +
#  hwdsb_ggplot_theme()
}
